modular code implementation
