+++ src/app.ts
@@ -1,347 +1,14 @@
-// Drag and Drop Interfaces
-interface Draggable {
-  dragStartHandler(event: DragEvent): void;
-  dragEndHandler(event: DragEvent): void;
-};
-
-interface DragTarget {
-  dragOverHandler(event: DragEvent): void;
-  dropHandler(event: DragEvent): void;
-  dragLeaveHandler(event: DragEvent): void;
-};
-
-// Project Class
-// reason for being a class to be able to instantiate it
-enum ProjectStatus { Active, Finished }
-class Project {
-  constructor(
-    public id: string,
-    public title: string,
-    public description: string,
-    public people: number,
-    public status: ProjectStatus) {
-    
-  }
+/// <reference path="models/drag-and-drop.ts" />
+/// <reference path="models/project.ts" />
+/// <reference path="state/project.ts" />
+/// <reference path="utils/validation.ts" />
+/// <reference path="decorators/autobind.ts" />
+/// <reference path="components/base-component.ts" />
+/// <reference path="components/project-input.ts" />
+/// <reference path="components/project-item.ts" />
+/// <reference path="components/project-list.ts" />
+namespace App {
+  new ProjectInput();
+  new ProjectList('active');
+  new ProjectList('finished');
 }
-
-// Project State Management
-type Listener<T> = (items: T[]) => void;
-
-class State<T> {
-  protected listeners: Listener<T>[] = [];
-
-  addListener(listenerFn: Listener<T>) {
-    this.listeners.push(listenerFn);
-  }
-}
-class ProjectState extends State<Project> {
-  private projects: Project[] = [];
-  private static instance: ProjectState;
-
-  private constructor() {
-    super();
-  }
-
-  static getInstance() {
-    if (this.instance) {
-      return this.instance;
-    }
-    this.instance = new ProjectState();
-    return this.instance;
-  }
-  
-  addProject(title: string, description: string, numOfPeople: number) {
-    const newProject = new Project(
-      Math.random().toString(),
-      title,
-      description,
-      numOfPeople,
-      ProjectStatus.Active
-    );
-
-    this.projects.push(newProject);
-    for(const listenerFn of this.listeners) {
-      listenerFn(this.projects.slice());
-    }
-  }
-
-  moveProject(projectId: string, newStatus: ProjectStatus) {
-    const project = this.projects.find(prj => prj.id === projectId);
-    if (project && project.status !== newStatus) {
-      project.status = newStatus;
-      this.updateListeners();
-    }
-  }
-
-  private updateListeners() {
-    for (const listenerFn of this.listeners) {
-      listenerFn(this.projects.slice());
-    }
-  }
-}
-
-const projectState = ProjectState.getInstance();
-
-// Validation
-interface Validatable {
-  value: string | number;
-  required?: boolean;
-  minLength? : number;
-  maxLength? : number;
-  min?: number;
-  max?: number
-}
-
-function validate(validatableInput: Validatable) {
-  let isValid = true;
-  if (validatableInput.required) {
-    isValid = isValid && validatableInput.value.toString().trim().length !== 0;
-  }
-
-  if (validatableInput.minLength != null && typeof validatableInput.value === 'string') {
-    isValid = isValid && validatableInput.value.length >= validatableInput.minLength;
-  }
-
-  if (validatableInput.maxLength != null && typeof validatableInput.value === 'string') {
-    isValid = isValid && validatableInput.value.length <= validatableInput.maxLength;
-  }
-
-  if (validatableInput.min != null && typeof validatableInput.value === 'number') {
-    isValid = isValid && validatableInput.value >= validatableInput.min;
-  }
-
-  if (validatableInput.max != null && typeof validatableInput.value === 'number') {
-    isValid = isValid && validatableInput.value <= validatableInput.max;
-  }
-  return isValid;
-}
-
-// Autobind decorator
-function Autobind(_target: any, _mehtodName: string, descriptor: PropertyDescriptor) {
-  const originalMethod = descriptor.value;
-  const adjustedDescriptor: PropertyDescriptor = {
-    configurable: true,
-    get() {
-      const boundFn = originalMethod.bind(this);
-      return boundFn;
-    }
-  }
-  return adjustedDescriptor;
-}
-
-// Component Base class
-abstract class Component<T extends HTMLElement, U extends HTMLElement> {
-  templateElement: HTMLTemplateElement;
-  hostElement: T;
-  element: U;
-
-  constructor(
-    templateId: string,
-    hostElementId: string,
-    insertAtStart: boolean,
-    newElementId?: string
-  ) {
-    this.templateElement = document.getElementById(templateId)! as HTMLTemplateElement;
-    this.hostElement = document.getElementById(hostElementId)! as T;
-    const importedNode = document.importNode(this.templateElement.content, true);
-    this.element = importedNode.firstElementChild as U; // Document Fragment\
-    if(newElementId) this.element.id = newElementId;
-
-    this.attach(insertAtStart);
-  }
-
-  private attach(insertAtBeggining: boolean) {
-    this.hostElement.insertAdjacentElement(
-      insertAtBeggining ? 'afterbegin' : 'beforeend',
-      this.element
-    );
-  }
-
-  abstract configure(): void;
-  abstract renderContent(): void;
-}
-
-// ProjectItem Class
-class ProjectItem extends Component<HTMLUListElement, HTMLLIElement>
-  implements Draggable{
-  private project: Project;
-
-  get persons() {
-    if (this.project.people === 1 ) {
-      return '1 person';
-    } else {
-      return `${this.project.people} persons`;
-    }
-  }
- 
-  constructor(hostId: string, project: Project) {
-    super('single-project', hostId, false, project.id);
-    this.project = project;
-    this.configure();
-    this.renderContent();
-  }
-
-  @Autobind
-  dragStartHandler(event: DragEvent) {
-    event.dataTransfer!.setData('text/plain', this.project.id);
-    event.dataTransfer!.effectAllowed = 'move';
-  }
-
-  dragEndHandler(_: DragEvent) {
-    console.log('DragEnd');
-  }
-
-  configure() {
-    this.element.addEventListener('dragstart', this.dragStartHandler);
-    this.element.addEventListener('dragend', this.dragStartHandler);
-  };
-  renderContent() {
-    this.element.querySelector('h2')!.textContent = this.project.title;
-    this.element.querySelector('h3')!.textContent = this.persons + ' assigned';
-    this.element.querySelector('p')!.textContent = this.project.description;
-  };
-}
-
-// ProjectList Class
-class ProjectList extends Component<HTMLDivElement, HTMLElement> 
-  implements DragTarget {
-  assignedProjects: Project[];
-
-  constructor(private type: 'active' | 'finished') {
-    super('project-list', 'app', false, `${type}-projects`);
-    this.assignedProjects = [];
-
-    this.configure();
-    this.renderContent();
-  }
-
-  @Autobind
-  dragOverHandler(event: DragEvent){
-    if (event.dataTransfer && event.dataTransfer.types[0] === 'text/plain') {
-      event.preventDefault();
-      const listEl = this.element.querySelector('ul')!;
-      listEl.classList.add('droppable');
-    }
-  }
-
-  @Autobind
-  dropHandler(event: DragEvent) {
-    const prjId = event.dataTransfer!.getData('text/plain');
-    projectState.moveProject(prjId, this.type === 'active' ? ProjectStatus.Active : ProjectStatus.Finished);
-  }
-  
-  @Autobind
-  dragLeaveHandler(_: DragEvent){
-    const listEl = this.element.querySelector('ul')!;
-    listEl.classList.remove('droppable');
-  }
-
-  configure() {
-    this.element.addEventListener('dragover', this.dragOverHandler);
-    this.element.addEventListener('dragleave', this.dragLeaveHandler);
-    this.element.addEventListener('drop', this.dropHandler);
-
-    projectState.addListener((projects: Project[]) => {
-      const relevantProjects = projects.filter(prj => {
-        if (this.type === 'active') {
-          return prj.status === ProjectStatus.Active;
-        }
-        return prj.status === ProjectStatus.Finished;
-      });
-      this.assignedProjects = relevantProjects;
-      this.renderProjects();
-    });
-  }
-
-  renderContent() {
-    const listId = `${this.type}-project-list`;
-    this.element.querySelector('ul')!.id = listId;
-    this.element.querySelector('h2')!.textContent = this.type.toUpperCase() + ' PROJECTS';
-  }
-
-  private renderProjects() {
-    const listElement = document.getElementById(`${this.type}-project-list`)! as HTMLUListElement;
-    listElement.innerHTML = '';
-    for(const prjItem of this.assignedProjects) {
-      new ProjectItem(this.element.querySelector('ul')!.id, prjItem);
-    }
-  }
-}
-
-// ProjectInput class
-class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
-  titleInput: HTMLInputElement;
-  descriptionInput: HTMLInputElement;
-  peopleInput: HTMLInputElement;
-
-  constructor() {
-    super('project-input', 'app', true, 'user-input');
-    this.titleInput = this.element.querySelector('#title') as HTMLInputElement;
-    this.descriptionInput = this.element.querySelector('#description') as HTMLInputElement;
-    this.peopleInput = this.element.querySelector('#people') as HTMLInputElement;
-    this.configure();
-  }
-
-  configure() {
-    this.element.addEventListener('submit', this.submitHandler);
-  }
-
-  private gatherUserInput(): [string, string, number] | void {
-    const enteredTitle = this.titleInput.value;
-    const enteredDescription = this.descriptionInput.value;
-    const enteredPeople = this.peopleInput.value;
-
-    const titleValidatable: Validatable = {
-      value: enteredTitle,
-      required: true,
-    }
-
-    const descriptionValidatable: Validatable = {
-      value: enteredDescription,
-      required: true,
-      minLength: 5,
-    }
-
-    const peopleValidatable: Validatable = {
-      value: +enteredPeople,
-      required: true,
-      min: 1,
-      max: 5,
-    }
-
-    if(
-      !validate(titleValidatable) ||
-      !validate(descriptionValidatable) ||
-      !validate(peopleValidatable)
-    ) {
-      alert('Invalid input, please try again');
-      return;
-    } else {
-      return [enteredTitle, enteredDescription, +enteredPeople];
-    }
-  }
-
-  private clearInputs() {
-    this.titleInput.value = '';
-    this.descriptionInput.value = '';
-    this.peopleInput.value = '';
-  }
-
-  @Autobind
-  private submitHandler(event: Event) {
-    event.preventDefault();
-    const userInput = this.gatherUserInput();
-    if (Array.isArray(userInput)) {
-      // if this returns true we got the Tuple type
-      const [title, description, people] = userInput;
-      projectState.addProject(title, description, people);
-      this.clearInputs();
-    }
-  }
-
-  renderContent() {}
-}
-
-const prjInput = new ProjectInput();
-const activePrjList  = new ProjectList('active');
-const finishedPrjList  = new ProjectList('finished');
