+++ src/models/drag-and-drop.ts
@@ -1,0 +1,13 @@
+// Drag and Drop Interfaces
+namespace App {
+  export interface Draggable {
+    dragStartHandler(event: DragEvent): void;
+    dragEndHandler(event: DragEvent): void;
+  };
+
+  export interface DragTarget {
+    dragOverHandler(event: DragEvent): void;
+    dropHandler(event: DragEvent): void;
+    dragLeaveHandler(event: DragEvent): void;
+  };
+}
